Started the flexbox project!
I made a basic html structure and started working
on the header part. I'm figuring out the way I want
to divide the containers because there are multiple
ways to do it. I also should get some images on hand
to add to my website. This will be cool once I'm done.